aesthetics_eval_pkg/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aesthetics-eval"
version = "0.1.0"
description = "Week-1 CLI wrapper around Aesthetics-Toolbox (no web UI), JSON outputs, Celery tasks, Docker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas==2.2.2",
    "numpy==1.26.4",
    "PyYAML==6.0.2",
    "jsonschema==4.23.0",
    "Pillow==10.3.0",
    "scikit-image==0.22.0",
    "lpips==0.1.4",
    "celery==5.3.6",
    "redis==4.6.0",
    "loguru==0.7.2",
    "click==8.1.7"
]

[project.scripts]
aesthetics-eval = "aesthetics_eval.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["aesthetics_eval*"]
exclude = ["tests*", "docker*", "out*", "tools*", "scripts*", ".github*"]
